@@ -10,20 +10,21 @@
     "check": "svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.0.2",
-    "@tsconfig/svelte": "^3.0.0",
-    "svelte": "^3.55.1",
-    "svelte-check": "^2.10.3",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "@tsconfig/svelte": "^5.0.4",
+    "svelte": "^5.1.13",
+    "svelte-check": "^4.0.6",
     "ts-standard": "^12.0.2",
-    "tslib": "^2.5.0",
-    "typescript": "^4.9.3",
-    "vite": "^4.1.0"
+    "tslib": "^2.8.1",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.10"
   },
   "dependencies": {
     "algebrite": "^1.4.0",
     "decimal.js": "^10.4.3",
-    "earcut": "^2.2.4",
+    "earcut": "^3.0.0",
     "leader-line-new": "^1.1.9",
-    "mathjs": "^11.5.1"
+    "mathjs": "^13.2.1",
+    "scratchblocks": "^3.6.4"
   }
 }
