@@ -5,31 +5,31 @@
   import Title from '../ui/Title.svelte'
 
   /** Division par 60 étape 1*/
-  let spanMinutes1: HTMLSpanElement
-  let spanSeconds1: HTMLSpanElement
-  let isMinutes1 = false
-  let isSeconds1 = false
-  let typeFeedbackDivisionPar60Etape1: '' | 'tryAgain'
+  let spanMinutes1: HTMLSpanElement = $state()
+  let spanSeconds1: HTMLSpanElement = $state()
+  let isMinutes1 = $state(false)
+  let isSeconds1 = $state(false)
+  let typeFeedbackDivisionPar60Etape1: '' | 'tryAgain' = $state()
   /** Division par 60 étape 2*/
-  let spanHours2: HTMLSpanElement
-  let spanMinutes2: HTMLSpanElement
-  let spanSeconds2: HTMLSpanElement
+  let spanHours2: HTMLSpanElement = $state()
+  let spanMinutes2: HTMLSpanElement = $state()
+  let spanSeconds2: HTMLSpanElement = $state()
   let isSeconds2 = false
   let isMinutes2 = false
   let isHours2 = false
   /** Division par 3600 étape 1*/
-  let spanHours36001: HTMLSpanElement
-  let spanMinutes36001: HTMLSpanElement
-  let spanSeconds36001: HTMLSpanElement
-  let timeInSeconds = randint(3800, 9000)
-  let isHours36001 = false
-  let isMinutes36001 = false
-  let isSeconds36001 = false
-  let typeFeedbackDivisionPar3600Etape1: '' | 'tryAgain'
-  let typeFeedbackFinal: '' | 'success' | 'tryAgain' = ''
+  let spanHours36001: HTMLSpanElement = $state()
+  let spanMinutes36001: HTMLSpanElement = $state()
+  let spanSeconds36001: HTMLSpanElement = $state()
+  let timeInSeconds = $state(randint(3800, 9000))
+  let isHours36001 = $state(false)
+  let isMinutes36001 = $state(false)
+  let isSeconds36001 = $state(false)
+  let typeFeedbackDivisionPar3600Etape1: '' | 'tryAgain' = $state()
+  let typeFeedbackFinal: '' | 'success' | 'tryAgain' = $state('')
 
-  let operation1Selected: string
-  let operation2Selected: string
+  let operation1Selected: string = $state()
+  let operation2Selected: string = $state()
 
   function initChecksPartie1(): void {
     isMinutes1 = false
@@ -141,8 +141,8 @@
 
   <p>Quel calcul souhaites-tu faire ?</p>
 
-  <select bind:value={operation1Selected} on:change={initChecksPartie1}>
-    <option value="" />
+  <select bind:value={operation1Selected} onchange={initChecksPartie1}>
+    <option value=""></option>
     <option value="Multiplication">Multiplier {timeInSeconds} par 60</option>
     <option value="Multiplication">Multiplier {timeInSeconds} par 3 600</option>
     <option value="Division">Diviser {timeInSeconds} par 60</option>
@@ -162,18 +162,18 @@
         Il y a 60 secondes dans une minute donc on cherche combien de fois il y a 60 s dans {timeInSeconds} s pour trouver le nombre de minutes.
       </p>
 
-      <div class="operationPosee" />
+      <div class="operationPosee"></div>
       {@html Operation({ operande1: timeInSeconds, operande2: 60, type: 'division' })}
       <div>
-        Donc {timeInSeconds} s = <span contenteditable="true" class="reponse" bind:this={spanMinutes1} on:keyup={checkMinutes1} /> min
-        <span contenteditable="true" bind:this={spanSeconds1} class="reponse" on:keyup={checkSeconds1} /> s.
+        Donc {timeInSeconds} s = <span contenteditable="true" class="reponse" bind:this={spanMinutes1} onkeyup={checkMinutes1}></span> min
+        <span contenteditable="true" bind:this={spanSeconds1} class="reponse" onkeyup={checkSeconds1}></span> s.
       </div>
       <Feedback type={typeFeedbackDivisionPar60Etape1} />
     </div>
     {#if isMinutes1 && isSeconds1}
       <p>Le nombre de minutes dépasse 60, donc on va chercher le nombre d'heures. Quel calcul souhaites-tu faire à présent ?</p>
-      <select bind:value={operation2Selected} on:change={initChecksPartie2}>
-        <option value="" />
+      <select bind:value={operation2Selected} onchange={initChecksPartie2}>
+        <option value=""></option>
         <option value="Multiplication">Multiplier {timeInSeconds % 60} par 60</option>
         <option value="Multiplication">Multiplier {Math.floor(timeInSeconds / 60)} par 60</option>
         <option value="DivisionQuotient">Diviser {Math.floor(timeInSeconds / 60)} par 60</option>
@@ -184,18 +184,18 @@
           <p>
             Il y a 60 minutes dans une heure donc on cherche combien de fois il y a 60 min dans {Math.floor(timeInSeconds / 60)} min.
           </p>
-          <div class="operationPosee" />
+          <div class="operationPosee"></div>
           {@html Operation({ operande1: Math.floor(timeInSeconds / 60), operande2: 60, type: 'division' })}
           <div>
             <div>
-              Finalement {timeInSeconds} s = <span contenteditable="true" class="reponse" on:keyup={checkHours2} bind:this={spanHours2} /> h
-              <span contenteditable="true" class="reponse" on:keyup={checkMinutes2} bind:this={spanMinutes2} />
-              min <span contenteditable="true" class="reponse" on:keyup={checkSeconds2} bind:this={spanSeconds2} /> s.
+              Finalement {timeInSeconds} s = <span contenteditable="true" class="reponse" onkeyup={checkHours2} bind:this={spanHours2}></span> h
+              <span contenteditable="true" class="reponse" onkeyup={checkMinutes2} bind:this={spanMinutes2}></span>
+              min <span contenteditable="true" class="reponse" onkeyup={checkSeconds2} bind:this={spanSeconds2}></span> s.
             </div>
             <Feedback type={typeFeedbackFinal} />
             {#if typeFeedbackFinal === 'success'}
               <div>
-                <button on:click={restartExercice}>Nouvelle Question</button>
+                <button onclick={restartExercice}>Nouvelle Question</button>
               </div>
             {/if}
           </div>
@@ -212,12 +212,12 @@
         Il y a 3 600 secondes dans une heure donc on cherche combien de fois il y a 3 600 s dans {timeInSeconds} s pour connaître le nombre d'heures.
       </p>
 
-      <div class="operationPosee" />
+      <div class="operationPosee"></div>
       {@html Operation({ operande1: timeInSeconds, operande2: 3600, type: 'division' })}
       <div>
-        Donc {timeInSeconds} s = <span contenteditable="true" class="reponse" bind:this={spanHours36001} on:keyup={checkHours36001} /> h
-        <span contenteditable="true" bind:this={spanMinutes36001} class="reponse" on:keyup={checkMinutes36001} /> min
-        <span contenteditable="true" bind:this={spanSeconds36001} class="reponse" on:keyup={checkSeconds36001} /> s.
+        Donc {timeInSeconds} s = <span contenteditable="true" class="reponse" bind:this={spanHours36001} onkeyup={checkHours36001}></span> h
+        <span contenteditable="true" bind:this={spanMinutes36001} class="reponse" onkeyup={checkMinutes36001}></span> min
+        <span contenteditable="true" bind:this={spanSeconds36001} class="reponse" onkeyup={checkSeconds36001}></span> s.
       </div>
       <Feedback type={typeFeedbackDivisionPar3600Etape1} />
     </div>
@@ -226,8 +226,8 @@
         Le nombre de secondes dépasse 60, donc on va chercher à convertir {timeInSeconds % 3600} s en minutes et secondes. Quel calcul souhaites-tu
         faire à présent ?
       </p>
-      <select bind:value={operation2Selected} on:change={initChecksPartie2}>
-        <option value="" />
+      <select bind:value={operation2Selected} onchange={initChecksPartie2}>
+        <option value=""></option>
         <option value="Multiplication">Multiplier {timeInSeconds % 3600} par 60</option>
         <option value="Multiplication">Multiplier {timeInSeconds % 3600} par 3600</option>
         <option value="DivisionReste">Diviser {timeInSeconds % 3600} par 60</option>
@@ -239,18 +239,18 @@
             Le nombre de secondes dépasse 60, donc on va chercher le nombre de minutes.
             <br />Il y a 60 secondes dans une minute donc on cherche combien de fois il y a 60 s dans {timeInSeconds % 3600} s.
           </div>
-          <div class="operationPosee" />
+          <div class="operationPosee"></div>
           {@html Operation({ operande1: Math.floor(timeInSeconds % 3600), operande2: 60, type: 'division' })}
           <div>
             <div>
-              Finalement {timeInSeconds} s = <span contenteditable="true" class="reponse" on:keyup={checkHours2} bind:this={spanHours2} /> h
-              <span contenteditable="true" class="reponse" on:keyup={checkMinutes2} bind:this={spanMinutes2} />
-              min <span contenteditable="true" class="reponse" on:keyup={checkSeconds2} bind:this={spanSeconds2} /> s.
+              Finalement {timeInSeconds} s = <span contenteditable="true" class="reponse" onkeyup={checkHours2} bind:this={spanHours2}></span> h
+              <span contenteditable="true" class="reponse" onkeyup={checkMinutes2} bind:this={spanMinutes2}></span>
+              min <span contenteditable="true" class="reponse" onkeyup={checkSeconds2} bind:this={spanSeconds2}></span> s.
             </div>
             <Feedback type={typeFeedbackFinal} />
             {#if typeFeedbackFinal === 'success'}
               <div>
-                <button on:click={restartExercice}>Nouvelle Question</button>
+                <button onclick={restartExercice}>Nouvelle Question</button>
               </div>
             {/if}
           </div>
