@@ -1,4 +1,14 @@
-<h1><slot></slot></h1>
+<script lang="ts">
+    interface Props {
+        children?: import('svelte').Snippet;
+    }
+
+    let { children }: Props = $props();
+    const children_render = $derived(children);
+
+</script>
+
+<h1>{@render children_render?.()}</h1>
 
 <style>
     h1 {
