@@ -1,8 +1,10 @@
+import { mount } from 'svelte'
 import './app.css'
 import App from './App.svelte'
 
-const app = new App({
-  target: document.getElementById('app')
-})
 
-export default app
+const divApp = document.getElementById("app")
+if (divApp) {
+    divApp.innerHTML = ''
+    const app = mount(App, { target: divApp })
+}
